requirements.txt
Django~=3.1.4
djangorestframework~=3.12.2
celery~=5.0.5
django-filter~=2.4.0
pip~=20.2.3
amqp~=5.0.2
vine~=5.0.0
pytz~=2020.4
kombu~=5.0.2
billiard~=3.6.3.0
click~=7.1.2
sqlparse~=0.4.1
asgiref~=3.3.1
wcwidth~=0.2.5
six~=1.15.0
setuptools~=49.2.1
requests~=2.25.1
redis==3.4.1